refactor: custom hook useCheckbox
