+++ src/components/Section.js
@@ -1,20 +1,8 @@
-import React, { useReducer } from 'react'
+import React from 'react'
 import Button from './Button'
 import SectionTask from './SectionTask'
 import { useSections } from '../hooks/asana/useSections.js'
-
-const reducer = (state, action) => {
-	switch (action.type) {
-		case 'check': {
-			return state.concat([action.checkbox])
-		}
-		case 'uncheck': {
-			return state.filter(checkbox => checkbox !== action.checkbox)
-		}
-		default:
-			return state
-	}
-}
+import { useCheckbox } from '../hooks/useCheckbox.js'
 
 function Section({ workspaceGid, assigneeGid, projectGid, updateTaskGids }) {
 	const { isFetching, sections } = useSections({ projectGid })
@@ -22,17 +10,9 @@ function Section({ workspaceGid, assigneeGid, projectGid, updateTaskGids }) {
 		Object.assign(section, { key: section.gid })
 	)
 
-	const [state, dispatch] = useReducer(reducer, [], () => [])
-	const checkCheckbox = checkedCheckbox => {
-		dispatch({ type: 'check', checkbox: checkedCheckbox })
-	}
-	const uncheckCheckbox = uncheckedCheckbox => {
-		dispatch({ type: 'uncheck', checkbox: uncheckedCheckbox })
-	}
+	const { checkedCheckboxes, checkCheckbox, uncheckCheckbox } = useCheckbox()
 
-	const handleClick = () => {
-		updateTaskGids(state)
-	}
+	const handleClick = () => updateTaskGids(checkedCheckboxes)
 
 	return (
 		<>
@@ -47,7 +27,7 @@ function Section({ workspaceGid, assigneeGid, projectGid, updateTaskGids }) {
 							workspaceGid={workspaceGid}
 							assigneeGid={assigneeGid}
 							sectionGid={section.gid}
-							checkedCheckboxes={state}
+							checkedCheckboxes={checkedCheckboxes}
 							checkCheckbox={checkCheckbox}
 							uncheckCheckbox={uncheckCheckbox}
 						/>
